+++ src/basherServiceProvider.php
@@ -13,9 +13,10 @@ class basherServiceProvider extends ServiceProvider
      */
     protected $commands = [
         'fitluismacedo\basher\Commands\Clean',
-        'fitluismacedo\basher\Commands\Enviroment',
-        'fitluismacedo\basher\Commands\Push',
         'fitluismacedo\basher\Commands\Deploy',
+        'fitluismacedo\basher\Commands\Enviroment',
+        'fitluismacedo\basher\Commands\Generate',
+        'fitluismacedo\basher\Commands\Push',
         'fitluismacedo\basher\Commands\Revert',
         'fitluismacedo\basher\Commands\Tag',
     ];